+++ src/AnalyticsClientInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace KinetiStack\Sdk;
+
+use KinetiStack\Sdk\Client\AnalyticsClientInterface as BaseAnalyticsClientInterface;
+
+interface AnalyticsClientInterface extends BaseAnalyticsClientInterface
+{
+}