+++ src/Enum/AnalyticsGrouping.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace KinetiStack\Sdk\Enum;
+
+enum AnalyticsGrouping: string
+{
+    case DAY = 'day';
+    case WEEK = 'week';
+    case MONTH = 'month';
+}